Binance USDT-M Futures Class

Class that implements the API methods for the Binance USDT-M Futures API.

Super classes

binance::Binance -> binance::BinanceFutures -> BinanceFuturesUSDTM

Methods

Public methods

BinanceFuturesUSDTM$new()

BinanceFuturesUSDTM$mark_price()

BinanceFuturesUSDTM$funding_rate_history()

BinanceFuturesUSDTM$funding_info()

BinanceFuturesUSDTM$price_ticker_v2()

BinanceFuturesUSDTM$open_interest_stats()

BinanceFuturesUSDTM$top_ls_ratio_accounts()

BinanceFuturesUSDTM$top_ls_ratio_positions()

BinanceFuturesUSDTM$ls_ratio()

BinanceFuturesUSDTM$taker_buy_sell_volume()

BinanceFuturesUSDTM$bltv_kline()

BinanceFuturesUSDTM$composite_index_info()

BinanceFuturesUSDTM$multi_asset_index()

BinanceFuturesUSDTM$clone()

Inherited methods

binance::Binance$check_server_time()

binance::Binance$get_exchange_info()

binance::Binance$kline()

binance::Binance$old_trade_lookup()

binance::Binance$order_book()

binance::Binance$price_ticker()

binance::Binance$print()

binance::Binance$recent_trades_list()

binance::Binance$sign_in()

binance::Binance$test_connectivity()

binance::Binance$ticker_stats_24hr()

binance::BinanceFutures$aggregate_trades()

binance::BinanceFutures$basis()

binance::BinanceFutures$continuous_kline()

binance::BinanceFutures$index_constituents()

binance::BinanceFutures$index_price_klines()

binance::BinanceFutures$mark_price_klines()

binance::BinanceFutures$open_interest()

binance::BinanceFutures$order_book_ticker()

binance::BinanceFutures$premium_index_klines()

binance::BinanceFutures$quarterly_settlement_price()

Method new()

Usage

BinanceFuturesUSDTM$new(testnet = FALSE)

Arguments

testnet

(logical) whether to use the testnet. Default is FALSE. Mark Price

Mark Price and Funding Rate

Method mark_price()

Usage

BinanceFuturesUSDTM$mark_price(symbol = NULL)

Arguments

symbol

(character) the trading symbol.

Returns

(tibble) a tibble with the mark price and funding rate data. Get Funding Rate History

Get funding rate history. Rate Limit 500/5min/IP

Method funding_rate_history()

Usage

BinanceFuturesUSDTM$funding_rate_history(
  symbol = NULL,
  start_time = NULL,
  end_time = NULL,
  limit = 100
)

Arguments

symbol

(character) the trading symbol.

start_time

(numeric) the start time for the klines in epoch milliseconds.

end_time

(numeric) the end time for the klines in epoch milliseconds.

limit

(integer) the number of results to return. Default is 100; max is 1000.

Returns

(tibble) a tibble with the funding rate history data. Get Funding Rate Info

Query funding rate info for symbols that had FundingRateCap/ FundingRateFloor / fundingIntervalHours adjustment

Method funding_info()

Usage

BinanceFuturesUSDTM$funding_info()

Returns

(tibble) a tibble with the funding rate info. Symbol Price Ticker V2

Latest price for a symbol or symbols.

Method price_ticker_v2()

Usage

BinanceFuturesUSDTM$price_ticker_v2(symbol = NULL)

Arguments

symbol

(character) the trading symbol.

Returns

(tibble) a tibble with the price ticker data. Open Interest Statistics

Get open interest statistics for a specific symbol.

Method open_interest_stats()

Usage

BinanceFuturesUSDTM$open_interest_stats(
  symbol,
  period,
  limit = 30,
  start_time = NULL,
  end_time = NULL
)

Arguments

symbol

(character) the trading symbol.

period

(character) the period for the open interest statistics. Possible values: "5m", "15m", "30m", "1h", "2h", "4h", "6h", "12h", "1d"

limit

(integer) the number of results to return. Default is 30; max is 500.

start_time

(numeric) the start time for the open interest statistics in epoch milliseconds.

end_time

(numeric) the end time for the open interest statistics in epoch milliseconds.

Returns

(tibble) a tibble with the open interest statistics. Top Trader Long/Short Ratio (Accounts)

Get top trader long/short ratio (accounts) for a specific symbol.

Method top_ls_ratio_accounts()

Usage

BinanceFuturesUSDTM$top_ls_ratio_accounts(
  symbol,
  period,
  limit = 30,
  start_time = NULL,
  end_time = NULL
)

Arguments

symbol

(character) the trading symbol.

period

(character) the period for the top trader long/short ratio. Possible values: "5m", "15m", "30m", "1h", "2h", "4h", "6h", "12h", "1d"

limit

(integer) the number of results to return. Default is 30; max is 500.

start_time

(numeric) the start time for the top trader long/short ratio in epoch milliseconds.

end_time

(numeric) the end time for the top trader long/short ratio in epoch milliseconds.

Returns

(tibble) a tibble with the top trader long/short ratio (accounts) data. Top Trader Long/Short Ratio (Positions)

Get top trader long/short ratio (positions) for a specific symbol.

Method top_ls_ratio_positions()

Usage

BinanceFuturesUSDTM$top_ls_ratio_positions(
  symbol,
  period,
  limit = 30,
  start_time = NULL,
  end_time = NULL
)

Arguments

symbol

(character) the trading symbol.

period

(character) the period for the top trader long/short ratio. Possible values: "5m", "15m", "30m", "1h", "2h", "4h", "6h", "12h", "1d"

limit

(integer) the number of results to return. Default is 30; max is 500.

start_time

(numeric) the start time for the top trader long/short ratio in epoch milliseconds.

end_time

(numeric) the end time for the top trader long/short ratio in epoch milliseconds.

Returns

(tibble) a tibble with the top trader long/short ratio (positions) data. Long/Short Ratio

Get long/short ratio for a specific symbol.

Method ls_ratio()

Usage

BinanceFuturesUSDTM$ls_ratio(
  symbol,
  period,
  limit = 30,
  start_time = NULL,
  end_time = NULL
)

Arguments

symbol

(character) the trading symbol.

period

(character) the period for the long/short ratio. Possible values: "5m", "15m", "30m", "1h", "2h", "4h", "6h", "12h", "1d"

limit

(integer) the number of results to return. Default is 30; max is 500.

start_time

(numeric) the start time for the long/short ratio in epoch milliseconds.

end_time

(numeric) the end time for the long/short ratio in epoch milliseconds.

Returns

(tibble) a tibble with the long/short ratio data. Taker Buy/Sell Volume

Get taker buy/sell volume for a specific symbol.

Method taker_buy_sell_volume()

Usage

BinanceFuturesUSDTM$taker_buy_sell_volume(
  symbol,
  period,
  limit = 30,
  start_time = NULL,
  end_time = NULL
)

Arguments

symbol

(character) the trading symbol.

period

(character) the period for the taker buy/sell volume. Possible values: "5m", "15m", "30m", "1h", "2h", "4h", "6h", "12h", "1d"

limit

(integer) the number of results to return. Default is 30; max is 500.

start_time

(numeric) the start time for the taker buy/sell volume in epoch milliseconds.

end_time

(numeric) the end time for the taker buy/sell volume in epoch milliseconds.

Returns

(tibble) a tibble with the taker buy/sell volume data. Historical BLVT NAV Kline/Candlestick

The BLVT NAV system is based on Binance Futures.

Method bltv_kline()

Usage

BinanceFuturesUSDTM$bltv_kline(
  symbol,
  interval,
  start_time = NULL,
  end_time = NULL,
  limit = 500
)

Arguments

symbol

(character) the trading symbol.

interval

(character) the interval for the klines.

start_time

(numeric) the start time for the klines in epoch milliseconds.

end_time

(numeric) the end time for the klines in epoch milliseconds.

limit

(integer) the number of results to return. Default is 500; max is 1500.

Returns

(tibble) a tibble with the kline data. Composite Index Symbol Information

Get composite index symbol information.

Method composite_index_info()

Usage

BinanceFuturesUSDTM$composite_index_info(symbol = NULL)

Arguments

symbol

(character) the trading symbol.

Returns

(tibble) a tibble with the composite index symbol information. Multi-Assets Mode Asset Index

Get Asset index for Multi-Assets mode.

Method multi_asset_index()

Usage

BinanceFuturesUSDTM$multi_asset_index(symbol = NULL)

Arguments

symbol

(character) the trading symbol.

Returns

(tibble) a tibble with the asset index data.

Method clone()

The objects of this class are cloneable with this method.

Usage

BinanceFuturesUSDTM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.