Binance Options Class

Class that implements the API methods for the Binance Options API.

Super class

binance::Binance -> BinanceOptions

Methods

Public methods

BinanceOptions$new()

BinanceOptions$get_exchange_info()

BinanceOptions$price_ticker()

BinanceOptions$mark_price()

BinanceOptions$excerise_records()

BinanceOptions$open_interest()

BinanceOptions$clone()

Inherited methods

binance::Binance$check_server_time()

binance::Binance$kline()

binance::Binance$old_trade_lookup()

binance::Binance$order_book()

binance::Binance$print()

binance::Binance$recent_trades_list()

binance::Binance$sign_in()

binance::Binance$test_connectivity()

binance::Binance$ticker_stats_24hr()

Method new()

Usage

BinanceOptions$new(testnet = FALSE)

Arguments

testnet

(logical) whether to use the testnet. Default is FALSE. Exchange Information

Current exchange trading rules and symbol information.

Method get_exchange_info()

Usage

BinanceOptions$get_exchange_info()

Returns

(list) a list with option_contract, option_asset, and option_symbol data. Symbol Price Ticker Get spot index price for option underlying.

Method price_ticker()

Usage

BinanceOptions$price_ticker(underlying)

Arguments

underlying

(character) the underlying symbol.

Returns

(tibble) a tibble with the index price data. Option Mark Price

Option mark price and greek info.

Method mark_price()

Usage

BinanceOptions$mark_price(symbol = NULL)

Arguments

symbol

(character) the trading symbol.

Returns

(tibble) a tibble with the mark price and option greeks data. Historical Exercise Records

Get historical exercise records. REALISTIC_VALUE_STRICKEN -> Exercised. EXTRINSIC_VALUE_EXPIRED -> Expired OTM.

Method excerise_records()

Usage

BinanceOptions$excerise_records(
  underlying = NULL,
  start_time = NULL,
  end_time = NULL,
  limit = 100
)

Arguments

underlying

(character) the underlying symbol.

start_time

(numeric) the start time for the exercise records in epoch milliseconds.

end_time

(numeric) the end time for the exercise records in epoch milliseconds.

limit

(integer) the number of results to return. Default is 100; max is 100.

Returns

(tibble) a tibble with the exercise records. Open interest

Get open interest for specific underlying asset on specific expiration date.

Method open_interest()

Usage

BinanceOptions$open_interest(underlying_asset, expiration)

Arguments

underlying_asset

(character) the underlying asset.

expiration

(character) the expiration date.

Returns

(tibble) a tibble with the open interest data.

Method clone()

The objects of this class are cloneable with this method.

Usage

BinanceOptions$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.